Move unzip method to `Seek::Util` to avoid annoying Zip namespace clash
